--- src/electron/platform/android/test-configs/tab-stops/test-config.scss
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-.how-to-test-list {
-    padding-left: 1em;
-}
