+++ src/electron/views/tab-stops/virtual-keyboard-buttons.scss
