+++ src/electron/views/tab-stops/virtual-keyboard-view.scss
