+++ src/tests/unit/tests/electron/views/tab-stops/virtual-keyboard-view.test.tsx
@@ -2,12 +2,12 @@
 // Licensed under the MIT License.
 
 import { NarrowModeStatus } from 'DetailsView/components/narrow-mode-detector';
-import { VirtualKeyboardButtonsDeps } from 'electron/views/virtual-keyboard/virtual-keyboard-buttons';
+import { VirtualKeyboardButtonsDeps } from 'electron/views/tab-stops/virtual-keyboard-buttons';
 import {
     VirtualKeyboardView,
     VirtualKeyboardViewDeps,
     VirtualKeyboardViewProps,
-} from 'electron/views/virtual-keyboard/virtual-keyboard-view';
+} from 'electron/views/tab-stops/virtual-keyboard-view';
 import { shallow } from 'enzyme';
 import * as React from 'react';
